requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
websockets==10.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
